Conflicts:
	MobRecharge/src/main/java/com/freecharge/controllers/UserController.java
	MobRecharge/target/classes/META-INF/maven/com.freecharge/MobRecharge/pom.properties
	MobRecharge/target/classes/com/freecharge/controllers/UserController.class

+++ MobRecharge/src/main/java/com/freecharge/jwt/payload/response/JwtResponse.java
@@ -8,7 +8,7 @@ import com.freecharge.entities.Gender;
 public class JwtResponse {
 	private String token;
 	private String type = "Bearer";
-	private Long id;
+	private Integer id;
 	private String username;
 	private String email;
 	private String mobileNumber;
@@ -19,7 +19,7 @@ public class JwtResponse {
 	
 	private List<String> roles;
 
-	public JwtResponse(String accessToken, Long id, String username, String email, String mobileNumber, Date dob,Gender gender,Date createdDate,Date updatedDate, List<String> roles) {
+	public JwtResponse(String accessToken, Integer id, String username, String email, String mobileNumber, Date dob,Gender gender,Date createdDate,Date updatedDate, List<String> roles) {
 		this.token = accessToken;
 		this.id = id;
 		this.username = username;
@@ -48,11 +48,11 @@ public class JwtResponse {
 		this.type = tokenType;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
