+++ MobRecharge/src/main/java/com/freecharge/security/services/UserDetailsImpl.java
@@ -18,7 +18,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class UserDetailsImpl implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private Integer id;
 
 	private String username;
 
@@ -35,7 +35,7 @@ public class UserDetailsImpl implements UserDetails {
 
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl(Long id, String username, String email, String password,String mobileNumber,Date dob,Gender gender,Date createdDate,Date updatedDate,
+	public UserDetailsImpl(Integer id, String username, String email, String password,String mobileNumber,Date dob,Gender gender,Date createdDate,Date updatedDate,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.username = username;
@@ -72,7 +72,7 @@ public class UserDetailsImpl implements UserDetails {
 		return authorities;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
